pyproject.toml
[tool.poetry]
name = "ufc-fight-analyzer"
version = "0.1.0"
description = "ai fight analyzer"
authors = ["David Cueva <dscueva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mediapipe = "^0.10.20"
opencv-python = "^4.10.0.84"
numpy = "^1.24"
ultralytics = "^8.3.59"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
